API Move DropdownFilter into its own namespace, add empty constructor array and rename default DB table
